perf(hooks): migrate agentation-context → SessionStart (M104 PR-B)
Follow-up to the #1234 hook-cache audit, same shape as PR-A (#1825).

agentation-context lived inside prompt/unified-dispatcher with runOnce: true,
which means every UserPromptSubmit fired the dispatcher, which probed a
session-keyed flag-file to decide whether to re-emit the reminder. The
.mcp.json contents don't change mid-session, so this is wasted I/O and
wasted re-evaluation.

Move the hook to its own SessionStart entry (async, 5s) so the .mcp.json
read happens once at boot and the reminder pins to the cached system-prompt
prefix. The output-guard allow-list shipped in #1822 makes this safe.

Changes:
- Move src/hooks/src/prompt/agentation-context.ts → src/hooks/src/lifecycle/
  - outputPromptContext() → outputSessionStartContext()
  - HOOK header rewritten for SessionStart lifecycle
- Drop import + HOOKS-array entry from prompt/unified-dispatcher (now 5 hooks)
- Register lifecycle/agentation-context in entries/lifecycle.ts
- New SessionStart hooks.json entry (async: true, timeout: 5)
- Update hooks.json description: 129→130 global, 197→198 total
- Update manifests/ork.json description: 197 → 198 hooks
- Move test src/hooks/src/__tests__/prompt/ → __tests__/lifecycle/
  with import path, mock function, hook_event, envelope, log msg updates
- Drop stale vi.mock of prompt/agentation-context from
  once-flag-plugin-data.test.ts (dispatcher no longer imports it)
- Update e2e dispatcher hook listing comment
- Bump count assertions:
  - async-registry + dispatcher-registry-wiring: 85 → 86 (3 sites)
  - hooks-json-wiring description total: 197 → 198
  - split-bundles totalHooks: 229 → 230

Verified live via dispatcher (all 4 paths, real fixture):
- agentation enabled in .mcp.json  → SessionStart envelope w/ reminder
- agentation disabled: true        → silent OK
- .mcp.json present, no agentation → silent OK
- no .mcp.json file                → silent OK

8292 hook unit tests pass, build clean, typecheck clean, manifests clean,
14/14 security pass, count-hooks.sh confirms GLOBAL=130 TOTAL=198.
