+++ docs/site/content/docs/foundations/choosing-a-plugin.mdx
@@ -11,7 +11,7 @@ OrchestKit ships as a single plugin: **ork**.
 claude install orchestkit/ork
 ```
 
-No choices, no configuration required. All 109 skills, 37 agents, and 197 hooks install together.
+No choices, no configuration required. All 109 skills, 37 agents, and 198 hooks install together.
 
 ## What Loads When
