+++ docs/site/content/docs/foundations/overview.mdx
@@ -1,6 +1,6 @@
 ---
 title: What is OrchestKit?
-description: The complete AI development toolkit for Claude Code — 109 skills, 37 agents, 197 hooks working together.
+description: The complete AI development toolkit for Claude Code — 109 skills, 37 agents, 198 hooks working together.
 ---
 
 import { Card, Cards } from 'fumadocs-ui/components/card';
@@ -24,7 +24,7 @@ When you type a prompt, OrchestKit works in three layers:
 
 <Steps>
   <Step>
-    ### Hooks (197) fire automatically
+    ### Hooks (198) fire automatically
 
     Security gates, auto-lint, and memory injection run on every action — you never invoke them directly.
   </Step>
@@ -42,7 +42,7 @@ When you type a prompt, OrchestKit works in three layers:
 
 ## One Plugin
 
-OrchestKit ships as a single plugin: **ork** — 109 skills, 37 agents, 197 hooks.
+OrchestKit ships as a single plugin: **ork** — 109 skills, 37 agents, 198 hooks.
 
 ## Next Steps
