+++ docs/site/content/docs/getting-started/first-10-minutes.mdx
@@ -15,7 +15,7 @@ After installing, verify everything is working:
 /ork:doctor
 ```
 
-This runs diagnostics on all 109 skills, 37 agents, and 197 hooks. If anything is misconfigured, it tells you exactly what to fix.
+This runs diagnostics on all 109 skills, 37 agents, and 198 hooks. If anything is misconfigured, it tells you exactly what to fix.
 
 ## Step 2: Your First Command Skill
