+++ docs/site/content/docs/getting-started/installation.mdx
@@ -13,7 +13,7 @@ import { LazySetupWizard } from '@/components/lazy';
 claude install orchestkit/ork
 ```
 
-This installs all 109 skills, 37 agents, and 197 hooks.
+This installs all 109 skills, 37 agents, and 198 hooks.
 
 ## Interactive Setup
 
@@ -35,7 +35,7 @@ OrchestKit Health Check
 Plugin:     ork v7.43.0
 Skills:     107 loaded (27 commands, 80 reference)
 Agents:     37 registered
-Hooks:      197 active (129 global, 46 agent, 22 skill-scoped)
+Hooks:      198 active (130 global, 46 agent, 22 skill-scoped)
 Memory:     Graph ✓  Local ✓  CC Native ✓
 MCP:        context7 ✓  memory ✓  tavily ○  agentation ○
