+++ docs/site/content/docs/hooks/architecture.mdx
@@ -11,7 +11,7 @@ This page is for plugin developers who want to understand how the hook system wo
 
 ## The Bundle System
 
-OrchestKit compiles 197 hooks into **12 event-specific bundles** using esbuild. When Claude Code fires a hook, only the bundle for that event type is loaded -- not the entire codebase.
+OrchestKit compiles 198 hooks into **12 event-specific bundles** using esbuild. When Claude Code fires a hook, only the bundle for that event type is loaded -- not the entire codebase.
 
 ```
 hooks.json entry
