+++ docs/site/content/docs/hooks/overview.mdx
@@ -1,5 +1,5 @@
 ---
-title: "197 Hooks: What Fires When"
+title: "198 Hooks: What Fires When"
 description: "TypeScript functions that intercept every Claude Code lifecycle event -- blocking dangerous commands, injecting context, and syncing memory, all invisibly."
 ---
 
@@ -31,12 +31,12 @@ type HookFn = (input: HookInput) => HookResult | Promise<HookResult>;
 
 | Category | Count | Description |
 |----------|-------|-------------|
-| **Global hooks** | 129 | Fire on every matching event regardless of agent |
+| **Global hooks** | 130 | Fire on every matching event regardless of agent |
 | **Agent-scoped hooks** | 46 | Fire only when a specific agent is active |
 | **Skill-scoped hooks** | 22 | Fire only during specific skill operations |
-| **Total** | **197** | Registered in `hooks.json`, compiled into 12 split bundles |
+| **Total** | **198** | Registered in `hooks.json`, compiled into 12 split bundles |
 
-All 197 hooks are written in TypeScript, compiled via esbuild into ESM bundles, and executed by Node.js. Zero external dependencies in production.
+All 198 hooks are written in TypeScript, compiled via esbuild into ESM bundles, and executed by Node.js. Zero external dependencies in production.
 
 ## Hook Events: The Full Timeline
