+++ docs/site/content/docs/reference/hooks/index.mdx
@@ -1,11 +1,11 @@
 ---
 title: Hooks Reference
-description: "Complete reference for all 129 OrchestKit hooks across 27 event categories."
+description: "Complete reference for all 130 OrchestKit hooks across 27 event categories."
 ---
 
 # Hooks Reference
 
-OrchestKit includes **129 hook entries** across **27 lifecycle event categories**.
+OrchestKit includes **130 hook entries** across **27 lifecycle event categories**.
 
 | Category | Hooks | Description |
 |----------|-------|-------------|
@@ -14,7 +14,7 @@ OrchestKit includes **129 hook entries** across **27 lifecycle event categories*
 | [PostToolUseFailure](/docs/reference/hooks/post-tool-use-failure) | 2 | Hooks for `PostToolUseFailure` events |
 | [PermissionRequest](/docs/reference/hooks/permission-request) | 4 | Hooks for `PermissionRequest` events |
 | [UserPromptSubmit](/docs/reference/hooks/user-prompt-submit) | 5 | Hooks for `UserPromptSubmit` events |
-| [SessionStart](/docs/reference/hooks/session-start) | 13 | Hooks for `SessionStart` events |
+| [SessionStart](/docs/reference/hooks/session-start) | 14 | Hooks for `SessionStart` events |
 | [CwdChanged](/docs/reference/hooks/cwd-changed) | 2 | Hooks for `CwdChanged` events |
 | [FileChanged](/docs/reference/hooks/file-changed) | 2 | Hooks for `FileChanged` events |
 | [SessionEnd](/docs/reference/hooks/session-end) | 6 | Hooks for `SessionEnd` events |
