+++ docs/site/content/docs/reference/hooks/session-start.mdx
@@ -1,14 +1,15 @@
 ---
 title: "SessionStart"
-description: "Hooks triggered on SessionStart events (13 hooks)."
+description: "Hooks triggered on SessionStart events (14 hooks)."
 ---
 
 # SessionStart Hooks
 
-**13 hooks** registered for the `SessionStart` lifecycle event.
+**14 hooks** registered for the `SessionStart` lifecycle event.
 
 | Hook | Matcher | Behavior | Description |
 |------|---------|----------|-------------|
+| `node` | `*` | 🔇 Silent | — |
 | `node` | `*` | 🔇 Silent | — |
 | `node` | `*` | 🔇 Silent | — |
 | `node` | `*` | 🔇 Silent | — |
