+++ docs/site/content/docs/reference/index.mdx
@@ -1,6 +1,6 @@
 ---
 title: Reference
-description: "Complete reference for all 109 skills, 37 agents, and 197 hooks."
+description: "Complete reference for all 109 skills, 37 agents, and 198 hooks."
 full: true
 ---
 
@@ -12,7 +12,7 @@ Browse the complete OrchestKit reference. Each entry is auto-generated from sour
 <Cards>
   <Card title="109 Skills" href="/docs/reference/skills" description="Reusable knowledge modules — 27 commands + 80 auto-injected references" />
   <Card title="37 Agents" href="/docs/reference/agents" description="Specialized AI personas with curated tools and skills" />
-  <Card title="197 Hooks" href="/docs/reference/hooks" description="TypeScript lifecycle automation — security gates, memory sync, audit logging" />
+  <Card title="198 Hooks" href="/docs/reference/hooks" description="TypeScript lifecycle automation — security gates, memory sync, audit logging" />
 </Cards>
 
 ## Demo Compositions
