+++ docs/site/content/docs/reference/skills/doctor.mdx
@@ -870,7 +870,7 @@ MCP Servers: WARNING
 
 ## Overview
 
-OrchestKit uses 129 global hook entries across 15 event types, compiled into 12 bundles. This reference explains how to validate and troubleshoot hooks.
+OrchestKit uses 130 global hook entries across 15 event types, compiled into 12 bundles. This reference explains how to validate and troubleshoot hooks.
 
 ## Hook Architecture
