+++ docs/site/content/docs/troubleshooting/faq.mdx
@@ -23,7 +23,7 @@ npm install -g @anthropic-ai/claude-code@latest
 
 ### How do I install OrchestKit?
 
-OrchestKit is a single plugin: **ork** — 109 skills, 37 agents, 197 hooks. Install with `claude install orchestkit/ork`.
+OrchestKit is a single plugin: **ork** — 109 skills, 37 agents, 198 hooks. Install with `claude install orchestkit/ork`.
 
 ### Can I use OrchestKit with other Claude Code plugins?
 
@@ -103,7 +103,7 @@ Yes. Create a markdown file in `src/agents/your-agent.md` with YAML frontmatter.
 
 ### How many hooks are there?
 
-197 total: 129 global hooks, 46 agent-scoped hooks, and 22 skill-scoped hooks. They are compiled into 12 split bundles for faster loading.
+198 total: 130 global hooks, 46 agent-scoped hooks, and 22 skill-scoped hooks. They are compiled into 12 split bundles for faster loading.
 
 ### What are fire-and-forget hooks?
