+++ manifests/ork.json
@@ -1,7 +1,7 @@
 {
   "name": "ork",
   "version": "7.90.0",
-  "description": "The Complete AI Development Toolkit — 109 skills, 37 agents, 197 hooks for full-stack development.",
+  "description": "The Complete AI Development Toolkit — 109 skills, 37 agents, 198 hooks for full-stack development.",
   "author": {
     "name": "Yonatan Gross",
     "url": "https://github.com/yonatangross"
