+++ src/hooks/hooks.json
@@ -1,5 +1,5 @@
 {
-  "description": "OrchestKit hooks - 129 global + 46 agent-scoped + 22 skill-scoped = 197 total hooks (18 once:true context loaders). M119 #1815: lifecycle/rules-size-check — SessionStart async hook warns when .claude/rules/*.md exceeds 35k chars (12.5% safety margin under CC's 40k auto-load cliff). Operator-facing stderr signal; references CONTRIBUTING-SKILLS.md storage-patterns (#1816). M119 PR-2 (#1794 follow-up): prompt/worktree-advisory-consumer — UserPromptSubmit reads .claude/state/worktree-advisory-*.md written by worktree-lifecycle-logger on monorepo detect, injects sparsePaths nudge as additionalContext (CC doesn't consume it on WorktreeCreate event itself). M119 (#1795) / M104 PR-A: lifecycle/ask-fallback-injector — SessionStart async hook injects text-prompt reminder when ORK_ASK_FALLBACK=text is set; pins to cached prompt prefix (previously UserPromptSubmit, re-injected per turn). Webhook forwarder is standalone async on ALL 27 events (v7.29.0: decoupled from dispatchers — zero inline forwarder calls). M121 #1489: posttool metrics-dispatcher fans out webhook-forwarder+metrics-bridge+context-crossing-warn in-process on the catchall matcher (2 hooks.json entries collapsed into 1). M119 #1476: nudge-outcome-resolver added on SessionStart to close the pre-compact-prompt feedback loop. M126 #1543: posttool/secret-handler scans Read/Bash/Grep/Glob for API keys + private blocks via CC 2.1.121 updatedToolOutput; OFF by default, ORK_SECRET_HOOK=AUDIT|REDACT to enable. M130 #1487: lifecycle/cc-version-check warns when running CC drifts from feature matrix. M140 Bundle B (#1790, #1791): prompt/goal-tracker + stop/goal-tracker write /goal convergence telemetry to .claude/state/goal-history.jsonl; lifecycle/goal-budget-guard (SessionEnd) enforces ORK_GOAL_MAX_TURNS_PER_SESSION (default 30) / ORK_GOAL_MAX_TOKENS_PER_SESSION (default 250k) by writing .claude/state/goal-budget-tripped.json, which prompt/goal-tracker reads with continueOnBlock:true so the user can override by deleting the file. All timeouts in SECONDS per CC spec.",
+  "description": "OrchestKit hooks - 130 global + 46 agent-scoped + 22 skill-scoped = 198 total hooks (18 once:true context loaders). M119 #1815: lifecycle/rules-size-check — SessionStart async hook warns when .claude/rules/*.md exceeds 35k chars (12.5% safety margin under CC's 40k auto-load cliff). Operator-facing stderr signal; references CONTRIBUTING-SKILLS.md storage-patterns (#1816). M119 PR-2 (#1794 follow-up): prompt/worktree-advisory-consumer — UserPromptSubmit reads .claude/state/worktree-advisory-*.md written by worktree-lifecycle-logger on monorepo detect, injects sparsePaths nudge as additionalContext (CC doesn't consume it on WorktreeCreate event itself). M119 (#1795) / M104 PR-A: lifecycle/ask-fallback-injector — SessionStart async hook injects text-prompt reminder when ORK_ASK_FALLBACK=text is set; pins to cached prompt prefix (previously UserPromptSubmit, re-injected per turn). M104 PR-B (#638): lifecycle/agentation-context — SessionStart async hook injects Agentation MCP annotation reminder when configured in .mcp.json; pins to cached prompt prefix (previously inside prompt/unified-dispatcher with runOnce per-turn file-flag check). Webhook forwarder is standalone async on ALL 27 events (v7.29.0: decoupled from dispatchers — zero inline forwarder calls). M121 #1489: posttool metrics-dispatcher fans out webhook-forwarder+metrics-bridge+context-crossing-warn in-process on the catchall matcher (2 hooks.json entries collapsed into 1). M119 #1476: nudge-outcome-resolver added on SessionStart to close the pre-compact-prompt feedback loop. M126 #1543: posttool/secret-handler scans Read/Bash/Grep/Glob for API keys + private blocks via CC 2.1.121 updatedToolOutput; OFF by default, ORK_SECRET_HOOK=AUDIT|REDACT to enable. M130 #1487: lifecycle/cc-version-check warns when running CC drifts from feature matrix. M140 Bundle B (#1790, #1791): prompt/goal-tracker + stop/goal-tracker write /goal convergence telemetry to .claude/state/goal-history.jsonl; lifecycle/goal-budget-guard (SessionEnd) enforces ORK_GOAL_MAX_TURNS_PER_SESSION (default 30) / ORK_GOAL_MAX_TOKENS_PER_SESSION (default 250k) by writing .claude/state/goal-budget-tripped.json, which prompt/goal-tracker reads with continueOnBlock:true so the user can override by deleting the file. All timeouts in SECONDS per CC spec.",
   "hooks": {
     "PreToolUse": [
       {
@@ -653,6 +653,16 @@
             "args": [
               "${CLAUDE_PLUGIN_ROOT}/hooks/bin/run-hook.mjs",
               "lifecycle/ask-fallback-injector"
+            ]
+          },
+          {
+            "type": "command",
+            "command": "node",
+            "async": true,
+            "timeout": 5,
+            "args": [
+              "${CLAUDE_PLUGIN_ROOT}/hooks/bin/run-hook.mjs",
+              "lifecycle/agentation-context"
             ]
           },
           {
