+++ src/hooks/src/__tests__/integration/once-flag-plugin-data.test.ts
@@ -38,9 +38,8 @@ vi.mock('../../prompt/handoff-injector.js', () => ({
   }),
 }));
 
-vi.mock('../../prompt/agentation-context.js', () => ({
-  agentationContext: () => ({ continue: true, suppressOutput: true }),
-}));
+// agentation-context migrated to lifecycle/SessionStart in M104 PR-B; the
+// dispatcher no longer imports it, so no mock is needed here.
 
 vi.mock('../../prompt/context-exhaustion-warner.js', () => ({
   contextExhaustionWarner: () => ({ continue: true, suppressOutput: true }),
