+++ src/hooks/src/__tests__/lifecycle/agentation-context.test.ts
@@ -1,11 +1,9 @@
-// Generated by OrchestKit Claude Plugin
-// Created: 2026-02-17
-
 /**
- * Unit tests for agentation-context hook
- * Tests UserPromptSubmit hook that injects annotation reminder when agentation MCP is configured
+ * Unit tests for lifecycle/agentation-context (#638, M104 PR-B)
  *
- * Issue #638: Add prompt hook for annotation context injection.
+ * Migrated from UserPromptSubmit (inside prompt/unified-dispatcher, runOnce)
+ * → SessionStart so the reminder pins to the cached prompt prefix instead
+ * of being re-evaluated on every prompt turn.
  */
 
 import { describe, it, expect, beforeEach, vi } from 'vitest';
@@ -29,11 +27,11 @@ vi.mock('node:fs', async () => {
 
 import {
   outputSilentSuccess,
-  outputPromptContext,
+  outputSessionStartContext,
 } from '../../lib/common.js';
 import { existsSync, readFileSync } from 'node:fs';
 
-import { agentationContext } from '../../prompt/agentation-context.js';
+import { agentationContext } from '../../lifecycle/agentation-context.js';
 import { createTestContext } from '../fixtures/test-context.js';
 
 // =============================================================================
@@ -42,12 +40,11 @@ import { createTestContext } from '../fixtures/test-context.js';
 
 function createInput(overrides: Partial<HookInput> = {}): HookInput {
   return {
-    hook_event: 'UserPromptSubmit',
-    tool_name: 'UserPromptSubmit',
+    hook_event: 'SessionStart',
+    tool_name: '',
     session_id: 'test-session',
     project_dir: '/test/project',
     tool_input: {},
-    prompt: 'Hello world test prompt',
     ...overrides,
   };
 }
@@ -70,20 +67,20 @@ function makeMcpConfig(agentationConfig: Record<string, unknown> | null = {}): s
 // =============================================================================
 
 let testCtx: ReturnType<typeof createTestContext>;
-describe('prompt/agentation-context', () => {
+describe('lifecycle/agentation-context (M104 PR-B, SessionStart)', () => {
   const mockExistsSync = vi.mocked(existsSync);
   const mockReadFileSync = vi.mocked(readFileSync);
   const mockOutputSilentSuccess = vi.mocked(outputSilentSuccess);
-  const mockOutputPromptContext = vi.mocked(outputPromptContext);
+  const mockOutputSessionStartContext = vi.mocked(outputSessionStartContext);
   beforeEach(() => {
     testCtx = createTestContext();
     vi.clearAllMocks();
     mockOutputSilentSuccess.mockReturnValue({ continue: true, suppressOutput: true });
-    mockOutputPromptContext.mockImplementation((ctx: string) => ({
+    mockOutputSessionStartContext.mockImplementation((ctx: string) => ({
       continue: true,
       suppressOutput: true,
       hookSpecificOutput: {
-        hookEventName: 'UserPromptSubmit' as const,
+        hookEventName: 'SessionStart' as const,
         additionalContext: ctx,
       },
     }));
@@ -100,7 +97,7 @@ describe('prompt/agentation-context', () => {
 
       expect(result.continue).toBe(true);
       expect(result.suppressOutput).toBe(true);
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
       expect(testCtx.log).toHaveBeenCalledWith(
         'agentation-context',
         'Agentation MCP not configured, skipping'
@@ -118,7 +115,7 @@ describe('prompt/agentation-context', () => {
 
       const result = agentationContext(createInput(), testCtx);
 
-      expect(mockOutputPromptContext).toHaveBeenCalledWith(
+      expect(mockOutputSessionStartContext).toHaveBeenCalledWith(
         expect.stringContaining('agentation_get_all_pending')
       );
       expect(result.continue).toBe(true);
@@ -131,7 +128,7 @@ describe('prompt/agentation-context', () => {
 
       agentationContext(createInput(), testCtx);
 
-      expect(mockOutputPromptContext).toHaveBeenCalledWith(
+      expect(mockOutputSessionStartContext).toHaveBeenCalledWith(
         expect.stringContaining('Agentation')
       );
     });
@@ -147,7 +144,7 @@ describe('prompt/agentation-context', () => {
 
       agentationContext(createInput(), testCtx);
 
-      expect(mockOutputPromptContext).toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).toHaveBeenCalled();
     });
 
     it('should log that agentation was detected', () => {
@@ -158,7 +155,7 @@ describe('prompt/agentation-context', () => {
 
       expect(testCtx.log).toHaveBeenCalledWith(
         'agentation-context',
-        'Agentation MCP detected — injecting annotation reminder'
+        'Agentation MCP detected — injecting annotation reminder (SessionStart, cached)'
       );
     });
   });
@@ -174,7 +171,7 @@ describe('prompt/agentation-context', () => {
       const result = agentationContext(createInput(), testCtx);
 
       expect(result.continue).toBe(true);
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
     });
   });
 
@@ -191,7 +188,7 @@ describe('prompt/agentation-context', () => {
       const result = agentationContext(createInput(), testCtx);
 
       expect(result.continue).toBe(true);
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
     });
 
     it('should return silent success when mcpServers key is missing', () => {
@@ -201,7 +198,7 @@ describe('prompt/agentation-context', () => {
       const result = agentationContext(createInput(), testCtx);
 
       expect(result.continue).toBe(true);
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
     });
   });
 
@@ -216,7 +213,7 @@ describe('prompt/agentation-context', () => {
       const result = agentationContext(createInput(), testCtx);
 
       expect(result.continue).toBe(true);
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
     });
   });
 
@@ -234,7 +231,7 @@ describe('prompt/agentation-context', () => {
       agentationContext(createInput(), testCtx);
 
       expect(mockExistsSync).toHaveBeenCalledTimes(2);
-      expect(mockOutputPromptContext).toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).toHaveBeenCalled();
     });
 
     it('should stop checking after finding agentation in the first file', () => {
@@ -291,7 +288,7 @@ describe('prompt/agentation-context', () => {
       expect(result.suppressOutput).toBe(true);
       // The catch inside isAgentationConfigured skips the file,
       // so it falls through to "not configured"
-      expect(mockOutputPromptContext).not.toHaveBeenCalled();
+      expect(mockOutputSessionStartContext).not.toHaveBeenCalled();
     });
 
     it('should return silent success when existsSync throws', () => {
