+++ src/hooks/src/lifecycle/agentation-context.ts
@@ -1,0 +1,103 @@
+// Generated by OrchestKit Claude Plugin
+// Created: 2026-05-14
+
+/**
+ * Agentation Context — SessionStart Hook (#638, M104 PR-B)
+ *
+ * Injects a brief reminder to check for pending UI annotations when the
+ * Agentation MCP server is configured in .mcp.json.
+ *
+ * Originally a UserPromptSubmit hook (issue #638) registered inside the
+ * unified-dispatcher with runOnce: true — so it fired on every prompt but
+ * early-returned after the first turn via a file-flag check.
+ *
+ * Migrated to SessionStart in M104 PR-B (follow-up to the #1234 audit): the
+ * .mcp.json contents do not change mid-session, so reading them at session
+ * boot and pinning the reminder to the cached prompt prefix is strictly
+ * cheaper than re-checking the file-flag on every prompt. Allow-listed by
+ * hooks/bin/output-guard.mjs after #1822.
+ *
+ * Hook event: SessionStart (async, timeout: 5s)
+ *
+ * MIGRATION NOTE (CC 2.1.118+, M122 #1501): kept here as a reference for the
+ * next step. The `type: "mcp_tool"` hook dispatch type would replace this
+ * reminder injection with a direct MCP call:
+ *
+ *   {
+ *     "SessionStart": [{
+ *       "type": "mcp_tool",
+ *       "tool": "mcp__agentation__agentation_get_all_pending",
+ *       "input": {},
+ *       "outputMapping": "additionalContext",
+ *       "condition": "agentation_enabled"
+ *     }]
+ *   }
+ *
+ * Deferred until the user-base floor reliably gates CC ≥ 2.1.118.
+ */
+
+import type { HookInput, HookResult, HookContext } from '../types.js';
+import {
+  outputSilentSuccess,
+  outputSessionStartContext,
+} from '../lib/common.js';
+import { existsSync, readFileSync } from 'node:fs';
+import { join } from 'node:path';
+import { NOOP_CTX } from '../lib/context.js';
+
+const HOOK_NAME = 'agentation-context';
+
+const AGENTATION_CONTEXT = `[Agentation] Check agentation_get_all_pending for pending UI annotations before starting work.`;
+
+/**
+ * Check if agentation is configured and enabled in .mcp.json.
+ * Checks both project-level `.mcp.json` and `.claude/mcp.json`.
+ */
+function isAgentationConfigured(projectDir: string): boolean {
+  const candidates = [
+    join(projectDir, '.mcp.json'),
+    join(projectDir, '.claude', 'mcp.json'),
+  ];
+
+  for (const mcpPath of candidates) {
+    if (!existsSync(mcpPath)) continue;
+
+    try {
+      const raw = readFileSync(mcpPath, 'utf8');
+      const config = JSON.parse(raw);
+      const servers = config.mcpServers || {};
+      const agentation = servers.agentation;
+
+      if (agentation && agentation.disabled !== true) {
+        return true;
+      }
+    } catch {
+      // Malformed JSON or read error — skip this file
+    }
+  }
+
+  return false;
+}
+
+/**
+ * Agentation context hook — injects a reminder to check pending annotations.
+ *
+ * Fires once at SessionStart. If the agentation MCP server is not configured
+ * or is disabled, returns silent success with no context injection.
+ */
+export function agentationContext(input: HookInput, ctx: HookContext = NOOP_CTX): HookResult {
+  try {
+    const projectDir = input.project_dir || ctx.projectDir;
+
+    if (!isAgentationConfigured(projectDir)) {
+      ctx.log(HOOK_NAME, 'Agentation MCP not configured, skipping');
+      return outputSilentSuccess();
+    }
+
+    ctx.log(HOOK_NAME, 'Agentation MCP detected — injecting annotation reminder (SessionStart, cached)');
+    return outputSessionStartContext(AGENTATION_CONTEXT);
+  } catch (err) {
+    ctx.log(HOOK_NAME, `Error checking agentation config: ${err}`, 'warn');
+    return outputSilentSuccess();
+  }
+}
