--- src/hooks/src/prompt/agentation-context.ts
@@ -1,114 +1,0 @@
-// Generated by OrchestKit Claude Plugin
-// Created: 2026-02-17
-
-/**
- * Agentation Context - UserPromptSubmit Hook (once: true)
- *
- * Injects a brief reminder to check for pending UI annotations when the
- * Agentation MCP server is configured in .mcp.json.
- *
- * Issue #638: Add prompt hook for annotation context injection.
- *
- * CC 2.1.9 Compliant: Uses hookSpecificOutput.additionalContext.
- *
- * MIGRATION NOTE (CC 2.1.118+, M122 #1501):
- * The `type: "mcp_tool"` hook dispatch type lets a hook invoke an MCP tool
- * directly without spawning a subprocess. When the user base is fully on
- * CC ≥ 2.1.118, the reminder-injection here can be replaced with a direct
- * dispatch in hooks.json:
- *
- *   {
- *     "UserPromptSubmit": [{
- *       "type": "mcp_tool",
- *       "tool": "mcp__agentation__agentation_get_all_pending",
- *       "input": {},
- *       "outputMapping": "additionalContext",
- *       "condition": "agentation_enabled"
- *     }]
- *   }
- *
- * This eliminates the round-trip: hook injects reminder → model reads it →
- * model calls MCP. With direct dispatch, the MCP result is injected once.
- *
- * See `src/skills/chain-patterns/references/mcp-tool-hooks.md` for the full
- * pattern and tier compatibility notes. Migration is deferred to a follow-up
- * once the floor reliably gates 2.1.118.
- */
-
-import type { HookInput, HookResult , HookContext} from '../types.js';
-import {
-  outputSilentSuccess,
-  outputPromptContext,
-} from '../lib/common.js';
-import { existsSync, readFileSync } from 'node:fs';
-import { join } from 'node:path';
-import { NOOP_CTX } from '../lib/context.js';
-
-// =============================================================================
-// CONSTANTS
-// =============================================================================
-
-const HOOK_NAME = 'agentation-context';
-
-const AGENTATION_CONTEXT = `[Agentation] Check agentation_get_all_pending for pending UI annotations before starting work.`;
-
-// =============================================================================
-// HELPERS
-// =============================================================================
-
-/**
- * Check if agentation is configured and enabled in .mcp.json
- * Checks both project-level and home-directory .mcp.json files.
- */
-function isAgentationConfigured(projectDir: string): boolean {
-  const candidates = [
-    join(projectDir, '.mcp.json'),
-    join(projectDir, '.claude', 'mcp.json'),
-  ];
-
-  for (const mcpPath of candidates) {
-    if (!existsSync(mcpPath)) continue;
-
-    try {
-      const raw = readFileSync(mcpPath, 'utf8');
-      const config = JSON.parse(raw);
-      const servers = config.mcpServers || {};
-      const agentation = servers.agentation;
-
-      if (agentation && agentation.disabled !== true) {
-        return true;
-      }
-    } catch {
-      // Malformed JSON or read error — skip this file
-    }
-  }
-
-  return false;
-}
-
-// =============================================================================
-// MAIN HOOK
-// =============================================================================
-
-/**
- * Agentation context hook — injects a reminder to check pending annotations.
- *
- * Runs once per session. If the agentation MCP server is not configured,
- * returns silent success with no context injection.
- */
-export function agentationContext(input: HookInput, ctx: HookContext = NOOP_CTX): HookResult {
-  try {
-    const projectDir = input.project_dir || (ctx.projectDir);
-
-    if (!isAgentationConfigured(projectDir)) {
-      ctx.log(HOOK_NAME, 'Agentation MCP not configured, skipping');
-      return outputSilentSuccess();
-    }
-
-    ctx.log(HOOK_NAME, 'Agentation MCP detected — injecting annotation reminder');
-    return outputPromptContext(AGENTATION_CONTEXT);
-  } catch (err) {
-    ctx.log(HOOK_NAME, `Error checking agentation config: ${err}`, 'warn');
-    return outputSilentSuccess();
-  }
-}
